Add shared constants to model message headers (#119)
## Description

Since in some headers we need to use some kind of enum, I've created a
constants package that contain the enum in order to avoid having them
defined in multiple repos.
Removed the `lint-with-docker` and `generate-with-docker`, because I
think not needed and hard to keep aligned with the original counter
part.

## Changes Made

N/A

## Related Issues

N/A

## Checklist

- [x] I have used a PR title that is descriptive enough for a release
note.
- [ ] I have tested these changes locally.
- [ ] I have added appropriate tests or updated existing tests.
- [ ] I have tested these changes on a cluster [name of the cluster] /
customer [name of the customer]
- [x] I have added appropriate documentation or updated existing
documentation.

---------

Co-authored-by: Federico M. Facca <chicco785@users.noreply.github.com>
Co-authored-by: Federico M. Facca <federico.facca@zaphiro.ch>
Co-authored-by: Bot <bot@zaphiro.ch>
Co-authored-by: chicco785 <777218+chicco785@users.noreply.github.com>